Adding checkbox buttons, fixing bugs
I fixed a lot of bugs today regarding the rooms page, I also implement a lot of customizations to the modals, inputs, and a lot of stuffs like (Optional) label that will be visible only if the input is not required, i also add a page named testing as a component test page for me so I can mess around with a lot of stuff. I also add the pre-script concist of external javascript declarations in the head so I can easily do javascript stuffs. I also implement the theme system for cards so that I can change the color of the cards whenever I want it. I also set the default style for disabled button which is mostly on buttons with btn-secondary class so that I can't click on it at all. There's also a lot of bug fixing here and there.